Addiction of factory file

+++ Product.API.Project/Product.API.Project/Entities/Products.cs
@@ -4,7 +4,7 @@
     {
         public Guid Id { get; set; }
         public string? Name { get; set; }
-        public decimal UnitValue { get; set; }
+        public double UnitValue { get; set; }
         public string? Seller { get; set; }
         public DateTime CreatedAt { get; private set; }
         public DateTime UpdatedAt { get; set; }
